+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - BREAKING CHANGE: modifiers options are now stored inside the modifiers configurations, see #62
 - Full support for HTML5 shadow root and HTML5 components (#88) (thanks @nadiam84)
 - Improved flip detection strategy (#82) (thanks @nadiam84)
+- BREAKING CHANGE: boundariesElement has been deprecated in favor to new options of `preventOverflow` and `flip` modifiers (#75)
 
 ## v0.6.4
 - 30a781e: Remove necessity of remove polyfill on IE (thanks @rafaelverger)